pull3: Feature/user (#6)
* migrate database + handle authentication logic

* add env sample

* fix:sửa warning, thêm redis khi logout

* fix: fix warning sunlint

* feature:follow/unfollow

* feature:update + fix:sunlint

* feat:getProfile + refector

* refector:add UserResponseDto

* fix:fix warning data-source

* feat:favorite article

* fix:resolve pull3

---------

+++ hello-nest/src/users/users.service.ts
@@ -5,8 +5,8 @@ import {
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-
 import { User } from './entities/user.entity';
+import { UserResponseDto } from './dto/user-response.dto';
 
 @Injectable()
 export class UsersService {
@@ -15,11 +15,12 @@ export class UsersService {
     private readonly usersRepository: Repository<User>,
   ) {}
 
-  async findAll(): Promise<User[]> {
-    return this.usersRepository.find();
+  async findAll(): Promise<UserResponseDto[]> {
+    const users: User[] = await this.usersRepository.find();
+    return users.map((user) => new UserResponseDto(user));
   }
 
-  async findById(id: string): Promise<User> {
+  async findById(id: number): Promise<UserResponseDto> {
     const user = await this.usersRepository.findOne({
       where: { id },
     });
@@ -28,19 +29,23 @@ export class UsersService {
       throw new NotFoundException('User not found');
     }
 
-    return user;
+    return new UserResponseDto(user);
   }
 
   async findByEmail(email: string): Promise<User | null> {
-    return this.usersRepository.findOne({
+    const user = await this.usersRepository.findOne({
       where: { email },
     });
+
+    return user?? null;
   }
 
-  async findByUsername(username: string): Promise<User | null> {
-    return this.usersRepository.findOne({
+  async findByUsername(username: string): Promise<UserResponseDto | null> {
+    const user = await this.usersRepository.findOne({
       where: { username },
     });
+
+    return user ? new UserResponseDto(user) : null;
   }
 
   async create(userData: Partial<User>): Promise<User> {
@@ -61,17 +66,90 @@ export class UsersService {
     return this.usersRepository.save(user);
   }
 
-  async update(id: string, userData: Partial<User>): Promise<User> {
-    const user = await this.findById(id);
+  async update(id: number, userData: Partial<User>): Promise<User> {
+    const user = await this.usersRepository.findOne({
+      where: { id },
+    });
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
 
     Object.assign(user, userData);
 
     return this.usersRepository.save(user);
   }
 
-  async remove(id: string): Promise<void> {
-    const user = await this.findById(id);
+  async remove(id: number) {
+    const user = await this.usersRepository.findOne({
+      where: { id },
+    });
 
-    await this.usersRepository.remove(user);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    return this.usersRepository.remove(user);
+  }
+
+  async getProfile(userId: number){
+    const user = await this.usersRepository.findOne({
+  where: { id: userId },
+  relations: {
+    articles: true,
+    comments: true,
+    followers: true,
+    following: true,
+  },
+  select:{
+    articles:{
+      id:true,
+      title:true,
+      content:true,
+      createdAt:true,
+      updatedAt:true,
+    },
+    comments:{
+      id:true,
+      articleId:true,
+      comment:true,
+    },
+    followers:{
+      followerId:true,
+      followingId:false,
+    },
+    following:{
+      followerId:false,
+      followingId:true,
+    }
+  }
+});
+
+    return  user;
+  }
+
+  async updateAvatar(
+    userId: number,
+    file: any,
+  ) {
+    const user = await this.usersRepository.findOne({
+      where: { id: userId },
+    });
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    // Đường dẫn lưu trong database
+    const avatarUrl = `/uploads/avatars/${file.filename}`;
+
+    user.avatar = avatarUrl;
+
+    await this.usersRepository.save(user);
+
+    return {
+      message: 'Update avatar successfully',
+      avatar: avatarUrl,
+    };
   }
 }
